requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.4
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
scikit-learn==1.6.1
scipy==1.15.2
sniffio==1.3.1
starlette==0.46.1
threadpoolctl==3.6.0
typing_extensions==4.12.2
uvicorn==0.34.0
